% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.topolink.output.R
\name{read.topolink.output}
\alias{read.topolink.output}
\title{Reads a collection of topolink log files given a log list file and a path to the log collection}
\usage{
read.topolink.output(mode = "observed", loglistLocation,
  topolinkLogsDirectory)
}
\arguments{
\item{mode}{The topolink mode. Can be one of "all", "observed".}

\item{loglistLocation}{The path to a text file containing the names of the log files. can be generated by 'ls s> loglist.txt' on the appropriate directory.}

\item{topolinkLogsDirectory}{The path to the directory containing the topolink logs.}
}
\value{
A list of parsed topolink logs
}
\description{
Reads a collection of topolink log files given a log list file and a path to the log collection
}
